Factored out policy headers from common
